+++ api/src/routers/sections.ts
@@ -8,6 +8,10 @@ const getSectionsInputSchema = z.object({
   courseId: z.string().uuid()
 })
 
+const getSectionByCallNumberInputSchema = z.object({
+  callNumber: z.number()
+})
+
 export const sectionsRouter = t.router({
   getSections: t.procedure.input(getSectionsInputSchema).output(z.array(sectionSchema)).query(async ({ input }) => {
   const { courseId } = input;
@@ -23,6 +27,18 @@ export const sectionsRouter = t.router({
     });
     const course = courses[0];
     return course.sections;
+  }),
+  getSectionByCallNumber: t.procedure.input(getSectionByCallNumberInputSchema).output(sectionSchema).query(async ({ input }) => {
+  const { callNumber } = input;
+    // use this prisma query when actual data exists
+    //return await prisma.subject.findMany();
+    const sections = await prisma.section.findMany({
+        where: {
+            callNumber: callNumber
+        }
+    });
+    const section = sections[0];
+    return section;
   }),
   // get subject by Id
   // validate with Zod
